stfu: add bash/fish/zsh alias support via fork+subshell+exec
In order to run 'somecmd someargs' in a subshell without heavy string
sanitization necessarily accompanying traditional
    someshell -c 'somecmd someargs'
we match the user's shell to its script (fish has non-posix syntax).

Therefore the local/share/ directory is used for most of the files,
while the bin is now a symlink.

This is only compatible with Linux.